Some changes to get the sound synthesis part working.
Also added a python script to the plate directory to automate modal
sound precomputation (at least partially)

+++ src/tools/eigensolver/arpack-eigensolver.cpp
@@ -366,7 +366,7 @@ int main(int argc, char* argv[])
 
     IntArray                 ids( numEigv );
 
-    for ( int i = 0; i < ids.size(); i++ ) {
+    for ( int i = 0; i < (int)ids.size(); i++ ) {
         ids[ i ] = i;
     }
